requirements.txt
asgiref==3.3.0
Django==3.1.2
django-phone-field==1.8.1
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
